글 삭제, modal, 이미지 에러 해결

+++ src/main/java/com/sns/post/PostRestController.java
@@ -56,13 +56,22 @@ public class PostRestController {
 	@DeleteMapping("/delete")
 	public Map<String, Object> delete(
 			@RequestParam("postId") int postId
-			,@RequestParam("userId") int userId
-			) {
+			,HttpSession session) {
 		
 		Map<String, Object> result = new HashMap<>();
 		
-		postBO.deletePostByPostIdUserId(postId, userId);
+		Integer userId = (Integer)session.getAttribute("userId");
+		if (userId == null) {
+			result.put("code", 500);
+			result.put("errorMessage", "로그인을 다시 해주세요.");
+		
+			return result;
+		}
 
+		postBO.deletePostByPostIdUserId(postId, userId);
+		result.put("code", 1);
+		result.put("result", "성공");
+		
 		return result;
 	}
 }
